[SSH] enh: Cast SSH exit code from Java deprecated Integer to Int

+++ modules/gridscale-ssh/src/main/scala/fr/iscpif/gridscale/ssh/SSHJobService.scala
@@ -68,7 +68,7 @@ object SSHJobService {
     val cmd = session.exec(cde.toString)
     try {
       cmd.join
-      (cmd.getExitStatus, IOUtils.readFully(cmd.getInputStream).toString, IOUtils.readFully(cmd.getErrorStream).toString)
+      (cmd.getExitStatus.toInt, IOUtils.readFully(cmd.getInputStream).toString, IOUtils.readFully(cmd.getErrorStream).toString)
     } finally cmd.close
   }
